libraries/python/neural_hive_observability/pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neural-hive-observability"
version = "1.1.0"
description = "Biblioteca de observabilidade padronizada para Neural Hive-Mind"
readme = {text = "Biblioteca Python que facilita instrumentação OpenTelemetry consistente entre serviços do Neural Hive-Mind com tracing, métricas e logging estruturado.", content-type = "text/plain"}
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Neural Hive-Mind Team"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: System :: Monitoring",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.21.0",
    "opentelemetry-instrumentation>=0.42b0",
    "opentelemetry-instrumentation-grpc>=0.42b0",
    "grpcio>=1.59.0",
    "prometheus-client>=0.19.0",
    "structlog>=23.2.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["neural_hive_observability*"]
exclude = ["tests", "tests.*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.json"]
